fix page limitation and ceate query service #4356
